Updating tests to new features

+++ broker/plugins/kubejobs/plugin.py
@@ -149,8 +149,7 @@ class KubeJobsExecutor(GenericApplicationExecutor):
 
             print "Job finished"
 
-            #time.sleep(float(30))
-            time.sleep(float(10))
+            time.sleep(float(30))
 
             if self.enable_visualizer:
                 visualizer.stop_visualization(api.visualizer_url,
@@ -161,7 +160,6 @@ class KubeJobsExecutor(GenericApplicationExecutor):
 
             # delete redis resources
             time.sleep(float(30))
-            #time.sleep(float(30))
             if not self.get_application_state() == 'terminated':
                 self.k8s.delete_redis_resources(self.app_id)
